gitupper_backend/requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
autopep8==1.6.0
azure-core==1.21.1
azure-storage-blob==12.9.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.10
colorama==0.4.4
configparser==5.2.0
constantly==15.1.0
crayons==0.4.0
cryptography==36.0.1
cssselect==1.1.0
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.47.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-rest-auth==0.9.5
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
fake-useragent==1.1.1
h11==0.13.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.10.1
incremental==21.3.0
isodate==0.6.1
lxml==4.7.1
msgpack==1.0.3
msrest==0.6.21
oauthlib==3.2.0
outcome==1.1.0
parse==1.19.0
Pillow==9.0.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyee==8.2.2
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyppeteer==1.0.2
pyquery==1.4.3
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-html==0.10.0
requests-oauthlib==1.3.1
selenium==4.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2021.5
undetected-chromedriver==3.1.5.post4
Unidecode==1.3.3
urllib3==1.26.8
w3lib==1.22.0
webdriver-manager==3.5.2
websockets==10.1
whitenoise==5.3.0
wsproto==1.0.0
zipp==3.7.0
zope.interface==5.4.0
